# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.UserResponse:
    properties:
      message:
        type: string
      stats:
        $ref: '#/definitions/models.UserStats'
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  models.UpdateProfileRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.UserStats:
    properties:
      account_age_days:
        type: integer
      created_at:
        type: string
      email:
        type: string
      last_login_at:
        type: string
      login_count:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  utils.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/utils.APIError'
      message:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: JWT Authentication API with user registration and login
  title: Authentication API
  version: "1.0"
paths:
...
  /api/user/account:
    delete:
      description: |-
        Permanently delete the authenticated user's account and all associated data

        **Deletion Process:**
        1. Verify user authentication and ownership
        2. Remove all user-associated data (files, records, etc.)
        3. Permanently delete user account from database
        4. Invalidate all existing JWT tokens for the user

        **Warning:**
        - This action is irreversible
        - All user data will be permanently lost
        - All uploaded files will be deleted

        **Authentication Required:**
        - Valid JWT token must be provided in Authorization header
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted successfully
          schema:
            $ref: '#/definitions/controllers.UserResponse'
        "401":
          description: Unauthorized - invalid or missing JWT token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to delete account
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - User Management
  /api/user/change-password:
    post:
      consumes:
      - application/json
      description: |-
        Change the authenticated user's password with current password verification

        **Password Requirements:**
        - Current password must be provided for verification
        - New password must be at least 6 characters long
        - New password must be different from current password
        - New password will be securely hashed with bcrypt

        **Change Process:**
        1. Verify current password is correct
        2. Validate new password meets requirements
        3. Hash new password securely
        4. Update password in database
        5. Optionally invalidate existing JWT tokens

        **Authentication Required:**
        - Valid JWT token must be provided in Authorization header
      parameters:
      - description: Password change data
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/controllers.UserResponse'
        "400":
          description: Bad request - validation errors or password requirements not
            met
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing JWT token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden - current password is incorrect
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to change password
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - User Management
  /api/user/profile:
    get:
      description: |-
        Retrieve the authenticated user's profile information including username, email, and account details

        **Profile Information:**
        - User ID and username
        - Email address
        - Account creation and last update timestamps
        - Account status and verification details

        **Authentication Required:**
        - Valid JWT token must be provided in Authorization header
        - Token format: "Bearer {your-jwt-token}"
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully
          schema:
            $ref: '#/definitions/controllers.UserResponse'
        "401":
          description: Unauthorized - invalid or missing JWT token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User profile not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to retrieve profile
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: |-
        Update the authenticated user's profile information such as email address

        **Updatable Fields:**
        - Email address (must be valid format and unique)
        - Additional profile fields as supported

        **Update Process:**
        1. Validate new email format and uniqueness
        2. Update user record in database
        3. Return updated profile information

        **Authentication Required:**
        - Valid JWT token must be provided in Authorization header
      parameters:
      - description: Profile update data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/controllers.UserResponse'
        "400":
          description: Bad request - invalid email format or validation errors
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing JWT token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Conflict - email already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to update profile
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User Management
  /api/user/stats:
    get:
      description: |-
        Retrieve comprehensive statistics about the authenticated user's account and activity

        **Statistics Include:**
        - Total number of uploaded files
        - Total storage space used
        - Account creation date and activity metrics
        - File type breakdown (images, documents, etc.)
        - Public vs private file counts

        **Use Cases:**
        - Dashboard displays
        - Storage usage monitoring
        - Account activity tracking
        - Usage analytics

        **Authentication Required:**
        - Valid JWT token must be provided in Authorization header
      produces:
      - application/json
      responses:
        "200":
          description: User statistics retrieved successfully
          schema:
            $ref: '#/definitions/controllers.UserResponse'
        "401":
          description: Unauthorized - invalid or missing JWT token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error - failed to retrieve statistics
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user statistics
      tags:
      - User Management
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"